MHD_add_response_header: Check on passed nullptr
The response argument is passed to `add_response_entry()` eventually
which does a check on NULL.  This was done without accessing struct
members of *response* in the past, however since 185f740e0684 ("allow
clients to override sanity check for content-length header") an access
to response->flags leads to a segfault.

This was spotted when building an app with libhttpserver which currently
might pass a nullptr to `MHD_add_response_header()`, see the bug report
over there for details.

Link: https://github.com/etr/libhttpserver/issues/255
Fixes: 185f740e0684 ("allow clients to override sanity check for content-length header")
Signed-off-by: Alexander Dahl <ada@thorsis.com>